+++ lib/core/bucket.js
@@ -19,6 +19,13 @@ module.exports = function bucket (self) {
     },
     deleteBucket: function(id, cb) {
 
+    },
+    /**
+      * makes a specific storage bucket public, uploading bucket key to bridge
+      * @param
+    */
+    makePublic: function(){
+
     }
   }
